server/requirement.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Django==3.0.4
Flask==1.1.1
gast==0.2.2
get==2019.4.13
idna==2.9
image==1.5.28
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.1
jsonpickle==1.3
MarkupSafe==1.1.1
numpy==1.18.2
opencv-python==4.2.0.32
Pillow==7.0.0
post==2019.4.13
public==2019.4.13
pytesseract==0.3.3
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.8
Werkzeug==1.0.0
wincertstore==0.2
